🛠️ [Chore]: View Components 분리
- CircleSlider 파일로 분리
